chore: move edit availability form to separate component
